implement functionalities

+++ storage.py
@@ -7,7 +7,7 @@ class Storage:
         self.conn = sqlite3.connect(db_name)
         self.cursor = self.conn.cursor()
         self.cursor.execute("CREATE TABLE IF NOT EXISTS users (username TEXT PRIMARY KEY, password_hash TEXT)")
-        self.cursor.execute("CREATE TABLE IF NOT EXISTS messages (id INTEGER PRIMARY KEY, sender TEXT, recipient TEXT, message TEXT)")
+        self.cursor.execute("CREATE TABLE IF NOT EXISTS messages (id INT PRIMARY KEY, sender TEXT, recipient TEXT, message TEXT, status TEXT)")
         self.conn.commit()
 
     def login_register_user(self, username, password):
@@ -30,29 +30,31 @@ class Storage:
     
     def list_accounts(self):
         self.cursor.execute("SELECT username FROM users")
-        return [row[0] for row in self.cursor.fetchall()]
+        response = {
+            'status': 'success',
+            'message': [row[0] for row in self.cursor.fetchall()]
+        }
+        return response
     
     def send_message(self, sender, recipient, message):
 
-        # check if sender and recipient exist
-        self.cursor.execute("SELECT username FROM users WHERE username=?", (sender,))
-        if not self.cursor.fetchone():
-            return {"status": "error", "message": "Sender does not exist"}
+        # check if recipient exist
         self.cursor.execute("SELECT username FROM users WHERE username=?", (recipient,))
         if not self.cursor.fetchone():
             return {"status": "error", "message": "Recipient does not exist"}
         
-        id = time.time()
-        id = f'{sender}-{recipient}-{id}'
+        id = int(time.time())
+        
         self.cursor.execute("INSERT INTO messages (id, sender, recipient, message, status) VALUES (?, ?, ?, ?, 'unread')", (id, sender, recipient, message))
         self.conn.commit()
         return {"status": "success"}
     
     def read_messages(self, username, limit=10):
         # read undelivered messages
-        self.cursor.execute("SELECT id, sender, message FROM messages WHERE sender=? and status='unread' ORDER BY id DESC LIMIT ?", (username, limit))
+        self.cursor.execute("SELECT id, sender, recipient, message FROM messages WHERE recipient=? and status='unread' ORDER BY id DESC LIMIT ?", (username, limit))
         
-        messages = [{"id": row[0], "sender": row[1], "message": row[2]} for row in self.cursor.fetchall()]
+        messages = [{"id": row[0], "sender": row[1], "message": row[3]} for row in self.cursor.fetchall()]
+        breakpoint()
         # mark messages as read
         for message in messages:
             self.cursor.execute("UPDATE messages SET status='read' WHERE id=?", (message["id"],))
@@ -60,17 +62,34 @@ class Storage:
         self.conn.commit()
 
         # if no undelivered messages, return a response
-        
-        return messages
+
+        if messages:
+            response = {
+                'status': 'success',
+                'messages': messages
+            }
+        else:
+            response = {
+                'status': 'error',
+                'messages': 'No messages unread'
+            }
+
+        return response
     
-    def delete_message(self, username, message_id):
-        self.cursor.execute("DELETE FROM messages WHERE id=? AND sender=?", (message_id, username))
+    def delete_message(self, username, recipient, message_id: str):
+        self.cursor.execute("DELETE FROM messages WHERE id=? AND sender=? AND recipient=?", (int(message_id), username, recipient))
         self.conn.commit()
         return {"status": "success"}
     
     def delete_account(self, username, password):
+        # check if the password is correct
+        self.cursor.execute("SELECT password_hash FROM users WHERE username=?", (username,))
+        user = self.cursor.fetchone()
+        if not bcrypt.checkpw(password.encode(), user[0]):
+            return {"status": "error", "message": "Invalid credentials"}
+
         # delete the account
-        self.cursor.execute("DELETE FROM users WHERE username=? AND password_hash=?", (username, bcrypt.hashpw(password.encode(), bcrypt.gensalt())))
+        self.cursor.execute("DELETE FROM users WHERE username=?", (username,))
         self.conn.commit()
         # delete all messages that sending to the account
         self.cursor.execute("DELETE FROM messages WHERE recipient=?", (username,))
